+++ packages/css-color-parser/test/test.mjs
@@ -1,5 +1,6 @@
 import './basic/basic.mjs';
 import './basic/color-function.mjs';
+import './basic/alpha-function.mjs';
 import './basic/color-mix-function-oklch.mjs';
 import './basic/color-mix-function.mjs';
 import './basic/contrast-color-function.mjs';