+++ plugins/postcss-progressive-custom-properties/scripts/contrast-color.mjs
@@ -4,7 +4,7 @@ export const contrastColorMatchers = [
 	{
 		'supports': 'contrast-color(red)',
 		'property': 'color',
-		'sniff': 'contrast-color',
+		'sniff': 'contrast-color(',
 		'matchers': [
 			matcherForValue('contrast-color($1)'),
 			matcherForValue('contrast-color($1)'),