+++ plugins/postcss-progressive-custom-properties/scripts/light-dark.mjs
@@ -4,7 +4,7 @@ export const lightDarkMatchers = [
 	{
 		'supports': 'light-dark(red, red)',
 		'property': 'color',
-		'sniff': 'light-dark',
+		'sniff': 'light-dark(',
 		'matchers': [
 			matcherForValue('light-dark($1,$2)'),
 		],