+++ plugins/postcss-progressive-custom-properties/src/index.ts
@@ -120,7 +120,6 @@ const creator: PluginCreator<null> = () => {
 						const parentClone = decl.parent.clone();
 						parentClone.removeAll();
 
-						// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 						parentClone.raws.before = '\n';
 
 						parentClone.append(decl.clone());